move vec_dot_f32_strided to buf_f32.rs
